feat: integrate between frontend and backend
