+++ src/usecase/student_pair.rs
@@ -1,6 +1,6 @@
 use std::sync::Arc;
 use crate::domain::student_pair::PairInfo;
-use crate::adapters::repository::StudentPairRepository;
+use crate::adapters::repository::student_pair::StudentPairRepository;
 use crate::infrastructure::models::StudentPair;
 use diesel::result::Error;
 use async_trait::async_trait;
